+++ CHANGELOG.md
@@ -6,7 +6,7 @@ Version 1.0.2
 
 _2016-11-24_
 
-* Use jsr305 at compiliation step.
+* Use `jsr305` at compilation step.
 
 Version 1.0.1 
 -------------